@@ -82,8 +82,8 @@
     "css-loader": "^3.4.2",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^5.6.1",
-    "eslint-config-ericgio": "^4.0.3",
+    "eslint": "^6.8.0",
+    "eslint-config-ericgio": "^5.0.1",
     "eslint-plugin-flowtype": "^4.6.0",
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
